[#48] Minor refactors from PR
